pyproject.toml
[project]
name = "techtracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "fastapi>=0.135.1",
    "fastmcp>=3.2.3",
    "feedparser>=6.0.12",
    "google-api-python-client>=2.191.0",
    "google-generativeai>=0.8.6",
    "isodate>=0.7.2",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "linkedin-api>=2.2.0",
    "lxml>=6.0.2",
    "python-dotenv>=1.2.2",
    "regex==2026.2.28",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "tavily-python==0.7.22",
    "tiktoken==0.12.0",
    "uvicorn>=0.41.0",
    "youtube-transcript-api>=1.2.4",
]
